Add onTeardown modifier.
https://github.com/Esri/arcgis-maps-sdk-swift-samples/pull/667#pullrequestreview-3048167758
