+++ Shared/Samples/Add items to portal/AddItemsToPortalView.swift
@@ -35,11 +35,6 @@ struct AddItemsToPortalView: View {
     /// The portal user when the portal is logged in.
     @State private var portalUser: PortalUser?
     
-    /// The action to run when the sample's teardown has completed.
-    /// 
-    /// This is needed to prevent the authentication in this sample from interfering with other samples.
-    @Environment(\.onTearDownCompleted) private var onTearDownCompleted
-    
     var body: some View {
         Group {
             if portalItems.isEmpty {
@@ -97,7 +92,7 @@ struct AddItemsToPortalView: View {
             apiKey = ArcGISEnvironment.apiKey
             ArcGISEnvironment.apiKey = nil
         }
-        .onDisappear {
+        .onTeardown {
             // Resetting the challenge handlers and clearing credentials here in
             // `onDisappear` so user is prompted to enter credentials every time
             // trying the sample. In real world applications, do these from
@@ -108,7 +103,7 @@ struct AddItemsToPortalView: View {
             
             // Sets the API key back to the original value.
             ArcGISEnvironment.apiKey = apiKey
-            Task { await signOut() }
+            await signOut()
         }
     }
     
@@ -182,8 +177,6 @@ struct AddItemsToPortalView: View {
     private func signOut() async {
         await ArcGISEnvironment.authenticationManager.revokeOAuthTokens()
         await ArcGISEnvironment.authenticationManager.clearCredentialStores()
-        
-        onTearDownCompleted()
     }
     
     /// Sets up new ArcGIS and Network credential stores that will be persisted in the keychain.
