+++ Shared/Samples/Analyze network with subnetwork trace/AnalyzeNetworkWithSubnetworkTraceView.swift
@@ -46,11 +46,6 @@ struct AnalyzeNetworkWithSubnetworkTraceView: View {
     /// The error shown in the error alert.
     @State private var error: Error?
     
-    /// The action to run when the sample's teardown has completed.
-    ///
-    /// This is needed to prevent the authentication in this sample from interfering with other samples.
-    @Environment(\.onTearDownCompleted) private var onTearDownCompleted
-    
     var body: some View {
         if !model.isSetUp {
             loadingView
@@ -96,9 +91,8 @@ struct AnalyzeNetworkWithSubnetworkTraceView: View {
             .toolbar {
                 ToolbarItemGroup(placement: .bottomBar) { toolbarItems }
             }
-            .onDisappear {
+            .onTeardown {
                 model.tearDown()
-                onTearDownCompleted()
             }
         }
     }
