+++ Shared/Samples/Display content of utility network container/DisplayContentOfUtilityNetworkContainerView.swift
@@ -43,11 +43,6 @@ struct DisplayContentOfUtilityNetworkContainerView: View {
     /// The current viewpoint of the map view.
     @State private var viewpoint = Viewpoint(latitude: 41.80, longitude: -88.16, scale: 4e3)
     
-    /// The action to run when the sample's teardown has completed.
-    ///
-    /// This is needed to prevent the authentication in this sample from interfering with other samples.
-    @Environment(\.onTearDownCompleted) private var onTearDownCompleted
-    
     var body: some View {
         MapViewReader { proxy in
             MapView(map: model.map, graphicsOverlays: [model.graphicsOverlay])
@@ -129,9 +124,8 @@ struct DisplayContentOfUtilityNetworkContainerView: View {
                         isShowingContainer = true
                     }
                 }
-                .onDisappear {
+                .onTeardown {
                     model.tearDown()
-                    onTearDownCompleted()
                 }
         }
     }
