+++ Shared/Samples/Edit with branch versioning/EditWithBranchVersioningView.swift
@@ -46,11 +46,6 @@ struct EditWithBranchVersioningView: View {
     /// The error shown in the error alert.
     @State private var error: Error?
     
-    /// The action to run when the sample's teardown has completed.
-    ///
-    /// This is needed to prevent the authentication in this sample from interfering with other samples.
-    @Environment(\.onTearDownCompleted) private var onTearDownCompleted
-    
     var body: some View {
         MapViewReader { mapViewProxy in
             MapView(map: model.map)
@@ -149,9 +144,8 @@ struct EditWithBranchVersioningView: View {
             }
         }
         .errorAlert(presentingError: $error)
-        .onDisappear {
+        .onTeardown {
             model.tearDown()
-            onTearDownCompleted()
         }
     }
