+++ Shared/Samples/Show utility associations/ShowUtilityAssociationsView.swift
@@ -34,11 +34,6 @@ struct ShowUtilityAssociationsView: View {
     /// The display scale of this environment.
     @Environment(\.displayScale) private var displayScale
     
-    /// The action to run when the sample's teardown has completed.
-    ///
-    /// This is needed to prevent the authentication in this sample from interfering with other samples.
-    @Environment(\.onTearDownCompleted) private var onTearDownCompleted
-    
     var body: some View {
         MapView(
             map: model.map,
@@ -65,9 +60,8 @@ struct ShowUtilityAssociationsView: View {
                 .shadow(radius: 3)
                 .padding()
         }
-        .onDisappear {
+        .onTeardown {
             model.tearDown()
-            onTearDownCompleted()
         }
     }
 }
