+++ Shared/Samples/Validate utility network topology/ValidateUtilityNetworkTopologyView.swift
@@ -34,11 +34,6 @@ struct ValidateUtilityNetworkTopologyView: View {
     /// The error shown in the error alert.
     @State private var error: Error?
     
-    /// The action to run when the sample's teardown has completed.
-    ///
-    /// This is needed to prevent the authentication in this sample from interfering with other samples.
-    @Environment(\.onTearDownCompleted) private var onTearDownCompleted
-    
     var body: some View {
         MapViewReader { mapViewProxy in
             MapView(map: model.map, graphicsOverlays: [model.graphicsOverlay])
@@ -139,9 +134,8 @@ struct ValidateUtilityNetworkTopologyView: View {
             }
         }
         .errorAlert(presentingError: $error)
-        .onDisappear {
+        .onTeardown {
             model.tearDown()
-            onTearDownCompleted()
         }
     }
 }
